Add RobotInfo per coordinator
