+++ nodeapp/nginx.conf
@@ -37,6 +37,9 @@ http {
         listen 12596;
         server_name robosats_client;
 
+        # add_header Access-Control-Allow-Headers "*";
+        # add_header Access-Control-Allow-Origin "*";
+
         location / {
             root /usr/src/robosats;
             try_files $uri $uri/ /basic.html;
